Thêm api review

+++ demo/src/main/java/com/doantotnghiep/demo/dto/request/user/AddBillRequest.java
@@ -1,11 +1,14 @@
 package com.doantotnghiep.demo.dto.request.user;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
+
 @Data
 public class AddBillRequest {
     private Integer status;
     private Integer discountPercent;
-    private Long priceTotal;
+    private BigDecimal priceTotal;
     private String couponName;
     private Long userId;
 }
