+++ demo/src/main/java/com/doantotnghiep/demo/dto/request/user/BuyRequest2.java
@@ -2,6 +2,7 @@ package com.doantotnghiep.demo.dto.request.user;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -9,7 +10,7 @@ public class BuyRequest2 {
     private Integer discountPersent;
     private String couponName;
     private Long userId;
-    private Long priceTotal;
+    private BigDecimal priceTotal;
 
     private List<AddBillProductDetailRequest> listBillProducts;
